Remove "templates" step from Hyprland configuration
Updating the configuration within the "files" folder won't work
with the files in "templates" folder because the ones copied from
"files" are overwritten with those in the "templates" folder.

Since the files in "files" folder are actively maintained and the
"templates" folder is left orphaned, nuke the "templates" folder
altogether.